+++ tests/test_isgri_plugin.py
@@ -5,6 +5,8 @@ import json
 import time
 import random
 
+from tests.test_basic import loop_ask, ask
+
 logger = logging.getLogger(__name__)
 
 default_params = dict(
@@ -47,14 +49,9 @@ def test_isgri_dummy(dispatcher_live_fixture, product_type):
     }
 
     logger.info("constructed server: %s", server)
-    c = requests.get(server + "/run_analysis",
-                      params=params)
-    logger.info("content: %s", c.text)
-    jdata = c.json()
+    jdata = ask(server, params, expected_query_status='done', expected_job_status='done', max_time_s=5)
     logger.info(list(jdata.keys()))
     logger.info(jdata)
-    assert c.status_code == 200
-    assert jdata['job_status'] == 'done'
 
 
 @pytest.mark.parametrize("selection", ["range", "280200470010.001"])
@@ -65,7 +62,7 @@ def test_isgri_image_no_pointings(dispatcher_live_fixture, selection):
     """
 
     server = dispatcher_live_fixture
-    print("constructed server:", server)
+    logger.info("constructed server: %s", server)
 
     if selection == "range":
         params = {
@@ -79,19 +76,11 @@ def test_isgri_image_no_pointings(dispatcher_live_fixture, selection):
             **default_params,
             'scw_list': selection
         }
-    # let's make the request public for simplicity
-    params.pop('token')
-    c = requests.get(server + "/run_analysis",
-                     params=params,
-                     )
 
-    jdata = c.json()
-
-    assert jdata["query_status"] == "failed"
+    jdata = ask(server, params, expected_query_status="failed", expected_job_status="failed", max_time_s=5)
     assert jdata["exit_status"]["debug_message"] == "{\"node\": \"dataanalysis.core.AnalysisException\", \"exception\": \"{}\", \"exception_kind\": \"handled\"}"
     assert jdata["exit_status"]["error_message"] == "AnalysisException:{}"
     assert jdata["exit_status"]["message"] == "failed: get dataserver products "
-    assert jdata["job_status"] == "failed"
 
 
 def test_isgri_image_fixed_done(dispatcher_live_fixture):
@@ -107,14 +96,7 @@ def test_isgri_image_fixed_done(dispatcher_live_fixture):
         'async_dispatcher': False,
     }
 
-    c = requests.get(server + "/run_analysis",
-                     params=params,
-                     )
-
-    jdata = c.json()
-
-    assert jdata["query_status"] == "done"
-
+    jdata = ask(server, params, expected_query_status="done")
     json.dump(jdata, open("jdata.json", "w"))
 
 
@@ -155,84 +137,4 @@ def test_isgri_image_random_emax(dispatcher_live_fixture):
         'E2_keV': emax,
     }
     jdata, tspent = loop_ask(server, params, max_time_s=5)
-
-
-def ask(server, params, expected_query_status, expected_job_status=None, max_time_s=2.0, expected_status_code=200):
-    t0 = time.time()
-    c=requests.get(server + "/run_analysis",
-                   params={**params},
-                  )
-    logger.info(f"\033[31m request took {time.time() - t0} seconds\033[0m")
-    t_spent = time.time() - t0
-    assert t_spent < max_time_s
-
-    logger.info("content: %s", c.text[:1000])
-    if len(c.text) > 1000:
-        print(".... (truncated)")
-
-    jdata=c.json()
-
-    if expected_status_code is not None:
-        assert c.status_code == expected_status_code
-
-    logger.info(list(jdata.keys()))
-
-    if expected_job_status is not None:
-        assert jdata["exit_status"]["job_status"] in expected_job_status
-
-    if expected_query_status is not None:
-        assert jdata["query_status"] in expected_query_status
-
-    return jdata
-
-
-def loop_ask(server, params, max_time_s=2.0):
-    jdata = ask(server,
-                params={
-                    **params,
-                    'async_dispatcher': True,
-                    'query_status': 'new'
-                },
-                expected_query_status=["submitted"],
-                max_time_s=max_time_s)
-
-    last_status = jdata["query_status"]
-
-    t0 = time.time()
-
-    tries_till_reset = 20
-
-    while True:
-        if tries_till_reset <= 0:
-            next_query_status = "ready"
-            print("\033[1;31;46mresetting query status to new, too long!\033[0m")
-            tries_till_reset = 20
-        else:
-            next_query_status = jdata['query_status']
-            tries_till_reset -= 1
-
-        jdata = ask(server,
-                    params={
-                        **params,
-                        "async_dispatcher": True,
-                        'query_status': next_query_status,
-                        'job_id': jdata['job_monitor']['job_id'],
-                        'session_id': jdata['session_id']
-                    },
-                    expected_query_status=["submitted", "done"],
-                    max_time_s=max_time_s,
-                    )
-
-        if jdata["query_status"] in ["ready", "done"]:
-            logger.info("query READY:", jdata["query_status"])
-            break
-
-        logger.info("query NOT-READY:", jdata["query_status"], jdata["job_monitor"])
-        logger.info("looping...")
-
-        time.sleep(5)
-
-    logger.info(f"\033[31m total request took {time.time() - t0} seconds\033[0m")
-
-    return jdata, time.time() - t0
